Moved files into gradle-compliant folder hierarchy
