+++ src/znoise/cpp/FBM.cpp
@@ -2,8 +2,8 @@
 // This file is part of ZNoise - a C++ noise library
 // For conditions of distribution and use, see LICENSE file
 
-#include "../include/CombinedNoiseBase.hpp"
-#include "../include/FBM.hpp"
+#include "CombinedNoiseBase.hpp"
+#include "FBM.hpp"
 
 FBM::FBM(NoiseBase& source): m_source(source)
 {
