+++ src/znoise/cpp/Perlin3D.cpp
@@ -2,8 +2,8 @@
 // This file is part of ZNoise - a C++ noise library
 // For conditions of distribution and use, see LICENSE file
 
-#include "../include/Perlin3D.hpp"
-#include "../include/NoiseTools.hpp"
+#include "Perlin3D.hpp"
+#include "NoiseTools.hpp"
 
 Perlin3D::Perlin3D()
 {
