+++ src/znoise/cpp/Perlin4D.cpp
@@ -2,8 +2,8 @@
 // This file is part of ZNoise - a C++ noise library
 // For conditions of distribution and use, see LICENSE file
 
-#include "../include/Perlin4D.hpp"
-#include "../include/NoiseTools.hpp"
+#include "Perlin4D.hpp"
+#include "NoiseTools.hpp"
 
 Perlin4D::Perlin4D()
 {
