+++ src/znoise/cpp/Simplex4D.cpp
