+++ src/znoise/headers/Enums.hpp
