+++ src/znoise/headers/HybridMultiFractal.hpp
