+++ src/znoise/headers/NoiseBase.hpp
